% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/high_dimensional_mean_comparison.R
\name{index_spliter}
\alias{index_spliter}
\title{A Cat Function}
\usage{
index_spliter(array, n_folds = 2)
}
\arguments{
\item{array}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\keyword{cats}
