% Generated by roxygen2 (4.0.2): do not edit by hand
\name{leapfrog}
\alias{leapfrog}
\title{Leapfrog steps}
\usage{
leapfrog(qp, dUdq, Minv, eps, L = 1L)
}
\arguments{
\item{qp}{data frame with columns \code{q} and \code{p} denoting the position and momentum of the particles respectively}

\item{dUdq}{derivative of potential energy}

\item{Minv}{inverse of the mass matrix}

\item{eps}{step-size}

\item{L}{number of steps to simulate}
}
\description{
Leapfrog steps
}
\examples{
nt <- 30
qp <- data.frame(q=0,p=1)
eps_gen <- function() 0.3
Minv = matrix(1,1,1)

dUdq <- function(q) q

for(i in 1:(nt-1))
  qp[i+1,] <- leapfrog(qp = qp[1,], dUdq = dUdq, Minv = Minv,eps = eps_gen(),L=i)

plot(qp$q,qp$p)
}

